% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilBlinkInterpolator.R
\name{pupilBlinkInterpolator}
\alias{pupilBlinkInterpolator}
\title{pupilBlinkInterpolator}
\usage{
pupilBlinkInterpolator(data, buffer = 5)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
see also ?zoo::na.approx
}
\examples{
add(1, 1)
add(10, 1)
}

