% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilCleaner.R
\name{pupilCleaner}
\alias{pupilCleaner}
\title{pupilCleaner}
\usage{
pupilCleaner(data, MinimumPupilSize = 2, MaximumPupilSize = 8,
  SDLimit = 3)
}
\arguments{
\item{data}{data.table from pupilPrepare}

\item{MinimumPupilSize}{remove values less than this value}

\item{MaximumPupilSize}{remove values greater than this value}

\item{SDLimit}{outlier limits in standard deviations}
}
\value{
NULL (data.table processed in place)
}
\description{
Attempts to clean the pupil data by removing and interpolating over extreme sample values
}
\details{
May help for datasets with extreme deviations
}
\examples{
pupilCleaner(pupilsamples)

}

