% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilNormalizer.R
\name{pupilNormalizer}
\alias{pupilNormalizer}
\title{pupilNormalizer}
\usage{
pupilNormalizer(data, baseline)
}
\arguments{
\item{data}{a data.table}

\item{baseline}{a TrialTime range to use for baseline}
}
\value{
processes the data.table in place
}
\description{
normalizes pupilsizes according to a specified baseline
(pupilsize - baselinesize)
}
\examples{
pupilNormalizer(data, c(100,1000))

}

