% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilParse.R
\docType{package}
\name{pupilParse-package}
\alias{pupilParse}
\alias{pupilParse-package}
\title{pupilParse}
\description{
helper functions for processing pupillometry data
}
\details{
this is a collections of functions that have been bundled together for easier access and sharing
the content of the present package is currently of alpha quality - use at your own risk
}
\examples{
data(pupilsamples)
pupilsamples = pupilPrepare(pupilsamples)


}

