% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{pupilPrepare}
\alias{pupilPrepare}
\title{pupilPrepare}
\usage{
pupilPrepare(data, subjectsColumn = "Subject", trialsColumn = "Trial",
  pupilSizeLeftColumn = "L.Mapped.Diameter..mm.",
  pupilSizeRightColumn = "R.Mapped.Diameter..mm.", samplingFrequency = 60,
  normalizeTrialDurations = TRUE)
}
\arguments{
\item{data}{a data.frame or data.table with pupil samples in rows}

\item{subjectsColumn}{name of column with subject identifiers}

\item{trialsColumn}{name of column with trial identifiers}

\item{pupilSizeLeftColumn}{name of column with left pupil size}

\item{pupilSizeRightColumn}{name of column with right pupil size}

\item{samplingFrequency}{sampling frequency of the eye-tracker (samples per second)}
}
\value{
a data table with specific columns
}
\description{
prepare a data.frame for processing with the functions in this package.
}
\examples{
data(pupilsamples)
pupilPrepare(pupilsamples, subjectsColumn = "Subject", trialsColumn = "Trial", pupilSizeLeftColumn = "L.Mapped.Diameter..mm.", pupilSizeRightColumn = "R.Mapped.Diameter..mm.", samplingFrequency = 60)


}

