% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pupilPreprocess}
\alias{pupilPreprocess}
\title{pupilPreprocess}
\usage{
pupilPreprocess(data)
}
\arguments{
\item{data}{a prepared data.table}
}
\description{
Not implemented.
}
\examples{
pupilPreprocess(data)
}

