% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilSmoother.R
\name{pupilSmoother}
\alias{pupilSmoother}
\title{pupilSmoother}
\usage{
pupilSmoother(...)
}
\arguments{
\item{...}{params for the smoothing method}

\item{data}{a prepared data.table}
}
\value{
processes the data.table in place
}
\description{
pupilSmoother
}
\examples{
pupilSmoother(data)

}

