% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilSmoother.R
\name{pupilSmoother_Hampel}
\alias{pupilSmoother_Hampel}
\title{pupilSmoother_Hampel}
\usage{
pupilSmoother_Hampel(data, Hampel_k = 15, Hampel_t0 = 1)
}
\arguments{
\item{data}{a prepared data.table}

\item{Hampel_k}{window length}

\item{Hampel_t0}{threshold}
}
\value{
processes the data.table in place
}
\description{
Hampel Filter
}
\details{
see ?pracma::hampel
}
\examples{
pupilSmoother_Hampel(data)

}

