% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilSmoother.R
\name{pupilSmoother_Loess}
\alias{pupilSmoother_Loess}
\title{pupilSmoother_Loess}
\usage{
pupilSmoother_Loess(data, Loess_degree = 2, Loess_span = 0.17)
}
\arguments{
\item{data}{prepared data.table}

\item{Loess_degree}{the degree of the polynomials to be used}

\item{Loess_span}{degree of smoothing}
}
\value{
processes the data.table in place
}
\description{
Local Polynomial Regression Fitting
}
\details{
see ??stats::loess
}
\examples{
pupilSmoother_Loess(data)

}

