% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PupilSmoother.R
\name{pupilSmoother_Lowess}
\alias{pupilSmoother_Lowess}
\title{pupilSmoother_Lowess}
\usage{
pupilSmoother_Lowess(data, Lowess_f = 0.1, Lowess_iter = 3L,
  Lowess_delta = 0.01)
}
\arguments{
\item{data}{a data.table}

\item{Lowess_f}{the smoother span}

\item{Lowess_iter}{the number of robustifying iterations}

\item{Lowess_delta}{defaults to 1/100th of the range of x}
}
\value{
processes the data.table in place
}
\description{
see ?stats::lowess
}
\examples{
pupilSmoother_Lowess(data)

}

